requirements.txt
fastapi
python-multipart
uvicorn
python-magic
